AOSCOS: target/loongarch: Add stubs for PTW feature
This is not a backport of the upstream commit because we don't want
backport the full software (TCG) PTW simulation.  We only want to
propagate the host PTW support status to the guest to avoid a severe
stability issue.  To ensure that, we even accounce PTW on a simulated
"LA464" or even "LA132" processor, sigh.  But the 10.2.0-rc1 release
will attempt to hide PTW for simulating them, deliberately make a merge
conflict to highlight the issue.

And I guess things will still wreck havoc if the guest kernel does not
support PTW.  Damn it.

Link: https://gitlab.com/qemu-project/qemu/-/commit/6447334a31bc
Signed-off-by: Xi Ruoyao <xry111@xry111.site>